soccial_token
Module accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is a Pubkey. This is useful for specifying accounts for a client.

Structs

BuyAndStakeTokens

Generated client accounts for BuyAndStakeTokens.

BuyTokensContext

Generated client accounts for BuyTokensContext.

ChangeUpdateAuthority

Generated client accounts for ChangeUpdateAuthority.

ContractToVaultContext

Generated client accounts for ContractToVaultContext.

CreateProposal

Generated client accounts for CreateProposal.

DepositTokensContext

Generated client accounts for DepositTokensContext.

EditVestingSchedule

Generated client accounts for EditVestingSchedule.

EmitSystemLog

Generated client accounts for EmitSystemLog.

FinalizeProposal

Generated client accounts for FinalizeProposal.

ImmutableVestingSchedule

Generated client accounts for ImmutableVestingSchedule.

InitializeEconomy

Generated client accounts for InitializeEconomy.

InitializeFoundersVesting

Generated client accounts for InitializeFoundersVesting.

InitializeSplToken

Generated client accounts for InitializeSplToken.

InitializeToken

Generated client accounts for InitializeToken.

ManageAirdrop

Generated client accounts for ManageAirdrop.

ManageContract

Generated client accounts for ManageContract.

ManageContractGovernance

Generated client accounts for ManageContractGovernance.

ManageGovernance

Generated client accounts for ManageGovernance.

ManageStaking

Generated client accounts for ManageStaking.

ManageUser

Generated client accounts for ManageUser.

ManageUserRemove

Generated client accounts for ManageUserRemove.

ManageVesting

Generated client accounts for ManageVesting.

ReinforceStake

Generated client accounts for ReinforceStake.

ReleaseStaked

Generated client accounts for ReleaseStaked.

ReleaseVestedTokens

Generated client accounts for ReleaseVestedTokens.

StakeTokens

Generated client accounts for StakeTokens.

TransferTokensContext

Generated client accounts for TransferTokensContext.

VaultDepositContext

Generated client accounts for VaultDepositContext.

VaultTransferContext

Generated client accounts for VaultTransferContext.

VaultWithdrawContext

Generated client accounts for VaultWithdrawContext.

VoteOnProposal

Generated client accounts for VoteOnProposal.

WithdrawStaked

Generated client accounts for WithdrawStaked.