soccial_token
Module instruction

An Anchor generated module containing the program’s set of instructions, where each method handler in the #[program] mod is associated with a struct defining the input arguments to the method. These should be used directly, when one wants to serialize Anchor instruction data, for example, when speciying instructions on a client.

Structs

AddAdmin

Instruction.

AddEarlyAdopter

Instruction.

AddFlag

Instruction.

AddStakingPlan

Instruction.

AddTokensToStake

Instruction.

Airdrop

Instruction.

AssignPermission

Instruction.

BuyAndStakeTokens

Instruction.

BuyTokens

Instruction.

CancelVestingSchedule

Instruction.

ClaimStakingRewards

Instruction.

ClaimVestedTokens

Instruction.

CreateProposal

Instruction.

CreateVestingSchedule

Instruction.

DepositTokens

Instruction.

DisableStakingPlan

Instruction.

EditStakingPlan

Instruction.

EmitSystemLog

Instruction.

FinalizeProposal

Instruction.

InitializeEconomy

Instruction.

InitializeFoundersVesting

Instruction.

InitializeSplToken

Instruction.

InitializeToken

Instruction.

MoveFromContractToVault

Instruction.

PauseContract

Instruction.

RemoveAdmin

Instruction.

RemoveFlag

Instruction.

ResumeContract

Instruction.

SetApiAuthority

Instruction.

SetVestingImmutable

Instruction.

StakeTokens

Instruction.

TransferBetweenVaults

Instruction.

TransferTokens

Instruction.

UnsignPermission

Instruction.

UpdateAirdropFee

Instruction.

UpdateGovernanceSettings

Instruction.

UpdateMetadataUri

Instruction.

UpdateRewardsFee

Instruction.

UpdateVersion

Instruction.

UpdateVestingSchedule

Instruction.

VaultDeposit

Instruction.

VaultWithdraw

Instruction.

Vote

Instruction.

WithdrawStakedTokens

Instruction.